Moved Calculations
calculations are moved to conversion.py
RPMtoVel is now in cm per sec
